+++ app/views/changeset_comments/timeout.html.erb
@@ -1,0 +1,1 @@
+<p><%= t '.sorry' %></p>
